+++ CHANGES.md
@@ -1,13 +1,15 @@
 
 # Revision History #
 
-## Version 1.5 -- Development snapshot C -- September 2017 ##
+## Version 1.5 -- Development snapshot C -- October 2017 ##
 
 This is a snapshot of ongoing development towards version of 1.5.  Some features might be incomplete or broken or not documented properly.
 
 ### New Features: ###
 
 - "kissutil" for troubleshooting a KISS TNC or interfacing to an application via files.
+
+- KISS "Set Hardware" command to report transmit queue length.
 
 
 ### Bugs Fixed: ###